The first attempt, I picked the wrong WPF platofrm. Should have been .NET Framework and not core.
Added the AddUserWindow.xaml and xml. Added the code to style them and get the functinality

Added also back into the project the BackendInterop class to call and link the dll to the frontend.

Modified the MainWindow xaml and xaml.cs files as needed

Debugging shows the window appears and works correctly